Move events tasks out of folder
