+++ augur/tasks/init/celery_app.py
@@ -29,7 +29,7 @@ start_tasks = ['augur.tasks.start_tasks',
 github_tasks = ['augur.tasks.github.contributors',
                 'augur.tasks.github.issues.tasks',
                 'augur.tasks.github.pull_requests.tasks',
-                'augur.tasks.github.events.tasks',
+                'augur.tasks.github.events',
                 'augur.tasks.github.messages.tasks',
                 'augur.tasks.github.facade_github.tasks',
                 'augur.tasks.github.releases.tasks',
